Add files via upload
It is used only if data doesn't follow normal distribution
used in image scaling
x -  xmin / xmax - xmin

Normalization is the process of organizing data in a database. This includes creating tables and establishing relationships between those tables according to rules designed both to protect the data and to make the database more flexible by eliminating redundancy and inconsistent dependency.